Move complex title constructor logic to action
